requirements.txt
django~=4.2.5
django-cors-headers~=4.2.0
mysqlclient~=2.2.0
python-dotenv~=1.0.0
requests~=2.31.0